part2/requirements.txt
torch==1.0.1
spacy==2.0.12
pyyaml==5.1
python-box==3.2.4
tqdm==4.31.1
ipdb==0.12
git+https://github.com/huggingface/pytorch-pretrained-BERT.git